--- a/avalanche/benchmarks/datasets/clear/__init__.py
+++ b/avalanche/benchmarks/datasets/clear/__init__.py
@@ -1,0 +1,2 @@
+from .clear import *
+from .clear10_data import *