--- a/avalanche/evaluation/metrics/detection_evaluators/lvis_evaluator.py
+++ b/avalanche/evaluation/metrics/detection_evaluators/lvis_evaluator.py
@@ -32,6 +32,7 @@ class LvisEvaluator:
 
     This mostly acts a wrapper around :class:`LVISEval` from the `lvis` library.
     """
+
     def __init__(self, lvis_gt: LVIS, iou_types: List[str]):
         assert isinstance(iou_types, (list, tuple))
         self.lvis_gt = lvis_gt
@@ -62,7 +63,8 @@ class LvisEvaluator:
                 self.predictions,
                 output if my_rank == 0 else None,
                 dst=0,
-                group=group)
+                group=group,
+            )
 
             return list(itertools.chain.from_iterable(output)), my_rank == 0
         else:
@@ -74,12 +76,12 @@ class LvisEvaluator:
             if max_dets_per_image is None:
                 max_dets_per_image = 300
 
-            eval_imgs = [lvis_res['image_id'] for lvis_res in all_preds]
+            eval_imgs = [lvis_res["image_id"] for lvis_res in all_preds]
 
             gt_subset = LvisEvaluator._make_lvis_subset(self.lvis_gt, eval_imgs)
 
             for iou_type in self.iou_types:
-                print('Evaluating for iou', iou_type)
+                print("Evaluating for iou", iou_type)
                 if iou_type == "segm":
                     # See:
                     # https://detectron2.readthedocs.io/en/latest/_modules/detectron2/evaluation/lvis_evaluation.html
@@ -89,8 +91,9 @@ class LvisEvaluator:
                 else:
                     lvis_results = all_preds
 
-                lvis_results = LVISResults(gt_subset, lvis_results,
-                                           max_dets=max_dets_per_image)
+                lvis_results = LVISResults(
+                    gt_subset, lvis_results, max_dets=max_dets_per_image
+                )
                 lvis_eval = LVISEval(gt_subset, lvis_results, iou_type)
                 lvis_eval.params.img_ids = list(set(eval_imgs))
                 lvis_eval.run()
@@ -127,9 +130,9 @@ class LvisEvaluator:
             scores = prediction["scores"].tolist()
             labels = prediction["labels"].tolist()
 
-            has_mask = 'mask' in prediction
-            has_bbox = 'boxes' in prediction
-            has_keypoint = 'keypoints' in prediction
+            has_mask = "mask" in prediction
+            has_bbox = "boxes" in prediction
+            has_keypoint = "keypoints" in prediction
 
             if has_bbox:
                 boxes = prediction["boxes"]
@@ -140,8 +143,11 @@ class LvisEvaluator:
                 masks = masks > 0.5
                 rles = [
                     mask_util.encode(
-                        np.array(mask[0, :, :, np.newaxis], dtype=np.uint8,
-                                 order="F"))[0] for mask in masks
+                        np.array(
+                            mask[0, :, :, np.newaxis], dtype=np.uint8, order="F"
+                        )
+                    )[0]
+                    for mask in masks
                 ]
                 for rle in rles:
                     rle["counts"] = rle["counts"].decode("utf-8")
@@ -161,10 +167,10 @@ class LvisEvaluator:
                     lvis_pred["bbox"] = boxes[pred_idx]
 
                 if has_mask:
-                    lvis_pred['segmentation'] = rles[pred_idx]
+                    lvis_pred["segmentation"] = rles[pred_idx]
 
                 if has_keypoint:
-                    lvis_pred['keypoints'] = keypoints[pred_idx]
+                    lvis_pred["keypoints"] = keypoints[pred_idx]
 
                 lvis_results.append(lvis_pred)
 
@@ -175,19 +181,19 @@ class LvisEvaluator:
         img_ids = set(img_ids)
 
         subset = dict()
-        subset['categories'] = list(lvis_gt.dataset["categories"])
+        subset["categories"] = list(lvis_gt.dataset["categories"])
 
         subset_imgs = []
         for img in lvis_gt.dataset["images"]:
             if img["id"] in img_ids:
                 subset_imgs.append(img)
-        subset['images'] = subset_imgs
+        subset["images"] = subset_imgs
 
         subset_anns = []
         for ann in lvis_gt.dataset["annotations"]:
             if ann["image_id"] in img_ids:
                 subset_anns.append(ann)
-        subset['annotations'] = subset_anns
+        subset["annotations"] = subset_anns
 
         return DictLVIS(subset)
 
@@ -207,9 +213,8 @@ class DictLVIS(LVIS):
         self.dataset = annotation_dict
 
         assert (
-                type(self.dataset) == dict
-        ), "Annotation file format {} not supported.".format(
-            type(self.dataset))
+            type(self.dataset) == dict
+        ), "Annotation file format {} not supported.".format(type(self.dataset))
         self._create_index()
 
 
@@ -218,6 +223,4 @@ def convert_to_xywh(boxes):
     return torch.stack((xmin, ymin, xmax - xmin, ymax - ymin), dim=1)
 
 
-__all__ = [
-    'LvisEvaluator'
-]
+__all__ = ["LvisEvaluator"]