--- a/avalanche/evaluation/metrics/loss.py
+++ b/avalanche/evaluation/metrics/loss.py
@@ -116,7 +116,7 @@ class LossPluginMetric(GenericPluginMetric[float]):
 
     def update(self, strategy):
         # task labels defined for each experience
-        if hasattr(strategy.experience, 'task_labels'):
+        if hasattr(strategy.experience, "task_labels"):
             task_labels = strategy.experience.task_labels
         else:
             task_labels = [0]  # add fixed task label if not available.