This PR updates the `OnlineSupervisedTemplate` and `InteractiveLogger`. As we discussed in our last meeting, all mini-experiences in online strategies now have three attributes assigned to them:
```
experience.is_online_exp
experience.online_exp_id
experience.online_exp_total
```

and these attributes are used to determine when to call model_adaptation, interactive logger updating and resetting, and potentially other conditional statement checking in the future.

I've tested it for NaiveOnline and NaiveOnline + Replay. For NaiveOnline, it works almost as expected; here are the profiler outputs:
ration for 

**Simple Online Naive Strategy (no Avalanche)**:  37.64 sec 

**Without interactive logger:**
OnlineNaiveStrategy:  84.01 sec
NaiveStrategy:  76.25

**With interactive logger:**
OnlineNaiveStrategy+IntLogger:  97.79 sec
NaiveStrategy+IntLogger:  111.2 sec

** the numbers can slightly change for different runs. We can say that the performances of Online and Batch-based strategies in Avalanche are similar.

With these changes, new strategies can be implemented with OnlineSupervisedTemplate, and the interactive logger works fine. However, the ReplayPlugin is slow, so the next step would be focusing on the performance of ReplayPlugin in online setups.