@AntonioCarta 2022-04-01
Maybe it's better to split this into two parts. First, we have to agree on what an `OnlineExperience` is, and how to define it. The current proposal:
```
experience.is_online_exp
experience.online_exp_id
experience.online_exp_total
```
breaks or reinvents a lot of features that I think we already have in place. It's ok if you want to change them, but it's not ok to duplicate the same logic with different names.

Some comments:
```
experience.is_online_exp --> instead of an attribute, can we do isinstance(OnlineExperience)? Otherwise, we may end up with several flags in the future just to check the type of an experience.
experience.online_exp_id --> isn't this experience.current_experience (i.e. the experience identifier)
experience.online_exp_total --> shouldn't this be a property of the stream?
```