@HamedHemati 2022-04-01
@AntonioCarta 
```
experience.is_online_exp --> instead of an attribute, can we do isinstance(OnlineExperience)? Otherwise, we may end up with several flags in the future just to check the type of an experience.

-> Then we would have to define OnlineExperience, which I think is not necessary. This just extends an experience to distinguish between the type (batch or online). I don't think we'll have other types?! it's just these two types.

experience.online_exp_id --> isn't this experience.current_experience (i.e. the experience identifier)

->not exactly, but a good idea, we can change it to : experience.current_experience 

experience.online_exp_total --> shouldn't this be a property of the stream?: 

->I don't think so, because the stream itself can have multiple experiences which are each split into individual mini-experiences. This total is related to the current actual experience, not the stream.
```