@HamedHemati 2022-04-01
@AntonioCarta I think the solution to online CL is very simple if we agree on changing the `ReplayPlugin`. 

By setting `n_epochs=1`, the strategy will become online (even if `n_batch!=1`). However, the biggest issue here is that the buffer gets updated only after each experience. So, if we create a ReplayPlugin that can also get updated after each iteration (can be chosen by the user), and change the `collate_fn` of the main data loader in a way that it combines the current batch with the sample from the buffer, I think we are done. If you think this is not correct, please let me know.