@HamedHemati 2022-04-01
> We already have RL, which is yet another scenario. We will add others in the future.

I think here the point is how many times we observe the data: just once or multiple times, not the problem type.

>There is no notion of mini-experiences in the final OCL stream. An experience in OCL is the single minibatch. This are generated from a larger experience, and we can use an ID to track this. It's also not clear to me what online_exp_total really is.

Maybe the naming is not ok, that's something that I also wanted to discuss. mini-experience is the previously called sub-experience that we had. I thought the word mini would be more meaningful as in online CL we have very small batches (either 1 or up to 10). So each mini experience is a part of a bigger experience that the model goes through. The `online_exp_total` I had to add only because of the interactive logger. You need to specify the total number of splits you make for the progress bar.