@HamedHemati 2022-04-01
I hope this visualization makes things more clear:
![Screenshot 2022-04-01 at 15 47 28](https://user-images.githubusercontent.com/11629528/161276440-ac625ee6-8076-4498-980e-e09dffef4a1d.png)