@ggraffieti 2022-04-14
Sorry to interrupt the discussion about this PR, but this is quite relevant and in line with the PR itself. 
The concept of splitting an experience into many "mini-experiences" is interesting and I'm currently dealing with it in the field of online CL. However, I don't think the current nomenclature of an experience that is split into many mini-experiences is totally correct since from the Avalanche paper an experience is _a set composed of one or multiple samples that can be used to update the model. This is often referred to in the literature as batch or task._ In the case of mini-experience, the experience cannot be used to update the model, since only a mini-experience is available. 
In our recent work we "flip" the relationship, and we called _experience_ the mini-experiences in the figure posted by @HamedHemati and _macro-experience_ the original big experience. 
IMO this is more coherent with the definition of experience as a chunk of data that can be used together to update the model. In this case, the macro-experience (another possible name is _season_) is just a collection of experiences that share some common properties. 
Maybe I should post this on the discussion panel since is more a nomenclature problem than an implementation one.