@vlomonaco 2022-04-14
Thanks @ggraffieti, we indeed discussed about this in the last avalanche meeting. In the end, the agent / strategy only sees a stream of experiences it cannot control, so there's no actual need to have a semantic distinction for the end users which remain the same discussed in the paper (@AntonioCarta correct me if I'm wrong).